ThreadPool dispatcher and cleanups
Change the default option to use a free worker if available, or to use the threadpool otherwise.
Immediate is too dangerous and error-prone to use as the default option.
IThreadPool abstraction, with cleaned up interface
MessageBus takes an IThreadPool constructor arg
Number of free workers in the thread pool is now a constructor-time
constant.