Rework RabbitMQ adaptor to improve functionality and fix bugs
Rabbit client requires a parameterless constructor to deserialize the payload, which Envelope<T> does not provide. Add a new RabbitEnvelope<T> to serve this purpose, with mappings back and forth
Move Local->Rabbit bindings to an extension method on SubscriberBuilder
We need to expose MessageBus on SubscriberBuilder, so we can get the module to do the magic
Change IMessageBus.Id to be a string, for more flexibility. Carry the change down to Envelope and other places where it is needed
Make local topic able to be a function of remote topic
Add builders for rabbit send/receive setups
Change Forwarding subscription to a subscriber reference, so we can select dispatch options on it
Remove old TODO notes and add new ones
Start updating some docs about Rabbit
s/Receiver/Consumer/g and interface segregation on IRabbitConsumerBuilder
Handle rebuilding queue after disconnect/reconnect
Add a TypeNameSerializer to simplify RabbitEnvelope<> exchange names
Fix a bug in MessageHandlerWorker exception reporting
misc fixes